Move getUTCDate, getDeciDate to databaseFunctions
refs #7137
